requirements.txt
fastapi
pydantic
jwt
passlib
pymongo
python-dotenv
pydantic
uvicorn
pydantic-settings
bcrypt
pyjwt
passlib[bcrypt]